`FlipSprites` in `HeroAnimComp` moved to header to avoid forward type declaration
- A bug sometimes not letting animation to be played fixed

- Reflection exposed to GunBase and RogueSpecial
