+++ CMakeLists.txt
@@ -59,7 +59,6 @@ add_executable(ETG main.cpp
         src/Engine/Reflection.h
         src/Core/GameClass.h
         src/Engine/UI/UIUtils.cpp
-        src/Core/Components/BaseAnimComp.cpp
 )
 
 # Easiest and fastest way is VCPKG
