+++ src/Characters/Components/HeroAnimComp.cpp
@@ -10,8 +10,6 @@ namespace ETG
     class RogueSpecial;
 }
 
-template sf::Vector2f ETG::HeroAnimComp::FlipSprites<ETG::RogueSpecial>(const Direction&, RogueSpecial&);
-
 namespace ETG
 {
     HeroAnimComp::HeroAnimComp()
@@ -19,37 +17,6 @@ namespace ETG
         HeroPtr = GameState::GetInstance().GetHero();
         IsGameObjectUISpecified = true;
         HeroAnimComp::SetAnimations();
-    }
-
-    template <typename... TObjects>
-    sf::Vector2f HeroAnimComp::FlipSprites(const Direction& currentDirection, TObjects&... objects)
-    {
-        if (!AnimManagerDict.contains(CurrentState)) return {8.f, 5.f};
-
-        const bool facingRight =
-            currentDirection == Direction::Right || currentDirection == Direction::FrontHandRight ||
-            currentDirection == Direction::BackDiagonalRight || currentDirection == Direction::BackHandRight;
-
-        //set horizontal scale to face the correct direction. Note that if abs not given, it will forever flip other direction in every tick 
-        HeroPtr->SetScale({
-            facingRight ? std::abs(HeroPtr->GetScale().x) : -std::abs(HeroPtr->GetScale().x),
-            HeroPtr->GetScale().y
-        });
-        
-        //lambda that applies the scale flip to an object. It'll only flip gun on Y axis according to hero's direction for now. 
-        auto flipObjectScale = [facingRight](auto& obj)
-        {
-            // obj.GetScale().y = flipFactor;
-            obj.SetScale({
-                obj.GetScale().x,
-                facingRight ? std::abs(obj.GetScale().y) : -std::abs(obj.GetScale().y)
-            });
-        };
-
-        //Apply the flip to all passed objects
-        (flipObjectScale(objects), ...);
-
-        return facingRight ? sf::Vector2f{8.f, 5.f} : sf::Vector2f{-7.f, 5.f};
     }
 
     void HeroAnimComp::SetAnimations()
@@ -105,17 +72,29 @@ namespace ETG
 
     void HeroAnimComp::Update()
     {
-        CurrentState = HeroPtr->CurrentHeroState;
+        AnimationKey newKey;
 
         if (HeroPtr->CurrentHeroState == HeroStateEnum::Dash)
-            CurrentAnimStateKey = InputComponent::GetDashDirectionEnum();
+            newKey = InputComponent::GetDashDirectionEnum();
 
         else if (HeroPtr->CurrentHeroState == HeroStateEnum::Run)
-            CurrentAnimStateKey = DirectionUtils::GetRunEnum(HeroPtr->CurrentDirection);
+            newKey = DirectionUtils::GetRunEnum(HeroPtr->CurrentDirection);
 
         else if (HeroPtr->CurrentHeroState == HeroStateEnum::Idle)
-            CurrentAnimStateKey = DirectionUtils::GetIdleDirectionEnum(HeroPtr->CurrentDirection);
+            newKey = DirectionUtils::GetIdleDirectionEnum(HeroPtr->CurrentDirection);
 
-        BaseAnimComp<HeroStateEnum>::Update(CurrentState, CurrentAnimStateKey);
+        if (newKey != CurrentAnimStateKey)
+        {
+            CurrentAnimStateKey = newKey;
+
+            //Restart the key
+            auto& animManager = AnimManagerDict[CurrentState];
+            if (animManager.AnimationDict.contains(CurrentAnimStateKey))
+            {
+                animManager.AnimationDict[CurrentAnimStateKey].Restart();
+            }
+        }
+
+        BaseAnimComp<HeroStateEnum>::Update(HeroPtr->CurrentHeroState, CurrentAnimStateKey);
     }
 }
